experiment: WIP NOT FASTER reuse threads for project discovery
This relaunches a task if one batch finished earlier than others and a
thread needs to call discoverProject. The downside is the extra graph
copy to start that new task and it doesn't really improve performences
overall.

A better but more complicate idea would be for a task that discovers a
thread is available to divide the remaining of its batch in two (if
there are more than a certain threshold of nodes in that remaining of
the batch) and submit a new task to handle one half of what's remaining.